ontology/requirements.txt
owlready2
tqdm
python-slugify
